requirements.txt
aioredis==1.2.0
aniso8601==7.0.0
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.7.0
boto==2.49.0
boto3==1.9.125
botocore==1.12.125
certifi==2018.11.29
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.4
defusedxml==0.5.0
Django==2.2.2
django-allauth==0.38.0
django-cleanup==3.2.0
django-cors-headers==2.2.0
django-countries==5.3.2
django-dbbackup==3.2.0
django-extra-fields==1.0.0
django-friendship==1.8.0
django-graphql-jwt==0.3.0
django-jet==1.0.8
django-rest-framework==0.1.0
django-rest-friendship==0.1.3
django-rest-passwordreset==1.0.0
django-ses==0.8.14
django-silk==3.0.1
django-sslserver==0.22
django-storages==1.7.1
django-webpush==0.2.7
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
dropbox==9.3.0
filetype==1.0.5
future==0.18.2
gprof2dot==2016.10.13
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.3.1
graphql-relay==2.0.1
hiredis==0.3.1
http-ece==1.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.19
jet-django==0.4.8
Jinja2==2.10
jmespath==0.9.4
lazy-object-proxy==1.4.1
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mkdocs==1.0.4
msgpack==0.6.0
msgpack-python==0.5.6
oauthlib==3.0.1
phonenumbers==8.10.14
Pillow==5.3.0
promise==2.3
psycopg2-binary==2.8.4
py-vapid==1.5.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.7.0
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-magic==0.4.15
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.7
pywebpush==1.6.0
PyYAML==3.13
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
rope==0.16.0
Rx==1.6.1
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
twilio==6.28.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.3.5
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
zope.interface==4.6.0
